Moved slices test to tests directory.
