+++ fits_reduce/util/slices.py
@@ -14,9 +14,3 @@ def slices_config(config):
     """
     return [slice_from_string(match.group()) for match in inside_brackets.finditer(config)]
 
-
-def test_slices():
-    assert slices_config('[0:10, 20:30], [30, 40:50], [13::], [::-1]') == [(slice(0, 10, None), slice(20, 30, None)),
-                                                                           (slice(None, 30, None), slice(40, 50, None)),
-                                                                           (slice(13, None, None),),
-                                                                           (slice(None, None, -1),)]
