+++ tests/test_config.py
@@ -2,7 +2,7 @@ import os
 from fits_reduce.util.config import FitsReduceConfig
 
 def test_config():
-    FitsReduceConfig('%s/../../conf.INI' % os.path.dirname(__file__), 'reducer')
+    FitsReduceConfig('%s/../conf.INI' % os.path.dirname(__file__), 'reducer')
     return True
